+++ static/css/account.css
@@ -1,0 +1,53 @@
+.account-header {
+    text-align: center;
+}
+.account-header>span{
+    margin: 50px;
+    cursor: pointer;
+}
+.account-main{
+    max-width: 600px;
+    width: 100%;
+    margin: 0 auto;
+}
+.order-number{
+    color: #448899;
+    font-weight: bold;
+}
+.order-info{
+    display: flex;
+    margin: 10px auto;
+}
+.img-box{
+    height: 197px;
+    border-radius: 10px;
+    overflow: hidden;
+}
+.order-img{
+    height: 100%;
+}
+.order-info-detail{
+    margin-left: 10px;
+    height: 197px;
+}
+@media(max-width:600px){
+    .account-header>span{
+        margin: 25px;
+    }
+    .order-info{
+        display: block;
+        width: 350px;
+    }
+    .img-box{
+        width: 100%;
+    }
+    .order-img{
+        width: 100%;
+        height: auto;
+    }
+}
+@media(max-width:400px){
+    .order-info{
+        width: 100%;
+    }
+}
