+++ static/css/booking.css
@@ -1,159 +1,10 @@
-body{
-    margin: 0;
-}
 a{
     text-decoration: none;
 }
-#headBox{
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    z-index: 99;
-    background-color: #ffffff;
-}
-#header{
-    background-color: #ffffff;
-    width: 1200px;
-    height: 54px;
-    display: flex;
-    margin: 0 auto;
-    
-}
-.header-left{
-    color: #448899;
-    font-size: 30px;
-    font-weight: bold;
-    flex: 50%;
-    margin: 1%;
-    display: flex;
-    align-items: center;
-    cursor: pointer;
-}
-.header-right{
-    display: flex;
-    justify-content: right;
-    align-items: center;
-    margin: 1%;
-    flex: 50%;
-    color: #666;
-}
-.header-rightItem{
-    margin-left: 20px;
-    cursor: pointer;
-}
-.header-rightItem>a{
-    color: #666;
-}
-.fakeHeadBox{
-    height: 54px;
-}
-.signMessage{
-    font-weight: bold;
-}
-.registerMessage{
-    font-weight: bold;
-}
-/* register */
-#login{
-    background-color: #ffffff;
-    width: 340px;
-    /* height: 275px; */
-    height: auto;
-    text-align: center;
-    position: fixed ;
-    display: none;
-    left: 0;
-    right: 0;
-    top: 80px;
-    margin: 0 auto;
-    border-radius: 5px;
-    border: none;
-    overflow: hidden;
-    z-index: 99;
-}
-#register{
-    background-color: #ffffff;
-    width: 340px;
-    /* height: 275px; */
-    height: auto;
-    text-align: center;
-    position: fixed ;
-    display: none;
-    left: 0;
-    right: 0;
-    top: 80px;
-    margin: 0 auto;
-    border-radius: 5px;
-    border: none;
-    overflow: hidden;
-    z-index: 99;
-}
-#loginWindowBackground{
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    min-height: 100%;
-    background: rgba(0,0,0,0.6);
-    z-index: 88;
-    cursor: pointer;
-    display: none;
-}
-.registerCloser{
-    position: absolute;
-    top: 20px;
-    right: 10px;
-    width: 20px;
-    height: 20px;
-    cursor: pointer;
-}
-.registerBackground{
-    background: linear-gradient(270deg, #337788 0%, #66AABB 100%);
-    width: 100%;
-    height: 10px;
-}
-.registerTitle{
-    color: #666666;
-    font-size: 24px;
-    font-weight: bold;
-    height: 44px;
-    margin-top: 10px;
-}
-.registerBox{
-    width: 95%;
-    margin: 0 auto;
-}
-.messageOK{
-    color: green;
-}
-.registerBoxEmail, .registerBoxPass, .loginButton, .registerButton, .newUserName, .newEmail, .newPassword{
-    max-width: 310px;
-    width: 90%;
-    height: 22px;
-    border-radius: 5px;
-    border: 1px solid #CCCCCC;
-    margin: 5px auto;
-    padding: 10px;
-}
-.loginButton ,.registerButton{
-    background-color: #448899;
-    color: #ffffff;
-    max-width: 312px; /*310+1border+1border*/
-    width: 100%;
-    font-size: 17px;
-    height: 42px;
-    padding: none;
-    border: none;
-    cursor: pointer;
-}
-.registerFooter{
-    height: 44px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
 /* 圖片+資訊 */
+#fakeBody{
+    display: none;
+}
 .fakeNav{
     background-color: #ffffff;
     width: 100%;
@@ -250,6 +101,9 @@ button{
     right: 10px;
     margin: 20px auto;
     cursor: pointer;
+}
+.fail-info{
+    color: red;
 }
 /* TapPay */
 .tpfield {
