+++ static/css/global.css
@@ -1,0 +1,188 @@
+body{
+    margin: 0;
+}
+#fakeHeader{
+    width: 100%;    
+    height: 54px;
+    margin: 0 auto;
+}
+.headerBackground{
+    background-color: #ffffff;
+    width: 100%;    
+    height: 54px;
+    position: fixed;
+    z-index: 21;
+    left: 50%;
+    right: 50%;
+    transform: translate(-50%, 0);
+}
+#header{
+    background-color: #ffffff;
+    width: 1200px;    
+    height: 54px;
+    display: flex;
+    position: fixed;
+    z-index: 22;
+    left: 50%;
+    right: 50%;
+    transform: translate(-50%, 0);
+}
+.header-left{
+    color: #448899;
+    font-size: 30px;
+    font-weight: bold;
+    flex: 50%;
+    margin: 1%;
+    display: flex;
+    align-items: center;
+}
+.header-right{
+    display: flex;
+    justify-content: right;
+    align-items: center;
+    margin: 1%;
+    flex: 50%;
+    color: #666;
+}
+.header-rightItem{
+    cursor: pointer;
+    margin: auto 3px;
+}
+.header-rightItem>a{
+    color: #666;
+}
+.signMessage{
+    font-weight: bold;
+}
+.registerMessage{
+    font-weight: bold;
+}
+/* register */
+#login{
+    background-color: #ffffff;
+    width: 340px;
+    height: auto;
+    text-align: center;
+    position: fixed ;
+    display: none;
+    left: 0;
+    right: 0;
+    top: 80px;
+    margin: 0 auto;
+    border-radius: 5px;
+    border: none;
+    overflow: hidden;
+    z-index: 99;
+}
+#register{
+    background-color: #ffffff;
+    width: 340px;
+    height: auto;
+    text-align: center;
+    position: fixed ;
+    display: none;
+    left: 0;
+    right: 0;
+    top: 80px;
+    margin: 0 auto;
+    border-radius: 5px;
+    border: none;
+    overflow: hidden;
+    z-index: 99;
+}
+#loginWindowBackground{
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    min-height: 100%;
+    background: rgba(0,0,0,0.6);
+    z-index: 88;
+    cursor: pointer;
+    display: none;
+}
+.registerCloser{
+    position: absolute;
+    top: 20px;
+    right: 10px;
+    width: 20px;
+    height: 20px;
+    cursor: pointer;
+}
+.registerBackground{
+    background: linear-gradient(270deg, #337788 0%, #66AABB 100%);
+    width: 100%;
+    height: 10px;
+}
+.registerTitle{
+    color: #666666;
+    font-size: 24px;
+    font-weight: bold;
+    height: 44px;
+    margin-top: 10px;
+}
+.registerBox{
+    width: 95%;
+    margin: 0 auto;
+}
+.messageOK{
+    color: green;
+}
+.registerBoxEmail, .registerBoxPass, .loginButton, .registerButton, .newUserName, .newEmail, .newPassword{
+    max-width: 310px;
+    width: 90%;
+    height: 22px;
+    border-radius: 5px;
+    border: 1px solid #CCCCCC;
+    margin: 5px auto;
+    padding: 10px;
+}
+.loginButton ,.registerButton{
+    background-color: #448899;
+    color: #ffffff;
+    max-width: 312px; /*310+1border+1border*/
+    width: 100%;
+    font-size: 17px;
+    height: 42px;
+    padding: none;
+    border: none;
+    cursor: pointer;
+}
+.registerFooter{
+    height: 44px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+footer{
+    color: #ffffff;
+    background-color: #757575;
+    height: 104px;
+    margin: 0 auto;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+@media (max-width:1200px){
+    #header{
+        width: 100%;
+    }
+    footer{
+        width: 100%;
+    }
+}
+@media(max-width:600px){
+    .header-rightItem{
+        margin-left: none;
+        line-height: 54px;
+    }
+}
+@media(max-width:370px){
+    #register{
+        width: 95%;
+    }
+    #login{
+        width: 95%;
+    }
+}
