--- static/index.css
@@ -1,355 +1,0 @@
-body{
-    margin: 0;
-}
-a{
-    text-decoration: none;
-}
-#header{
-    background-color: #ffffff;
-    width: 1200px;
-    height: 54px;
-    display: flex;
-    margin: 0 auto;
-}
-.header-left{
-    color: #448899;
-    font-size: 30px;
-    font-weight: bold;
-    flex: 50%;
-    margin: 1%;
-    display: flex;
-    align-items: center;
-}
-.header-right{
-    display: flex;
-    justify-content: right;
-    align-items: center;
-    margin: 1%;
-    flex: 50%;
-    color: #666;
-}
-.header-rightItem{
-    margin-left: 20px;
-    cursor: pointer;
-}
-.header-rightItem>a{
-    color: #666;
-}
-.signMessage{
-    font-weight: bold;
-}
-.registerMessage{
-    font-weight: bold;
-}
-/* register */
-#login{
-    background-color: #ffffff;
-    width: 340px;
-    /* height: 275px; */
-    height: auto;
-    text-align: center;
-    position: fixed ;
-    display: none;
-    left: 0;
-    right: 0;
-    top: 80px;
-    margin: 0 auto;
-    border-radius: 5px;
-    border: none;
-    overflow: hidden;
-    z-index: 99;
-}
-#register{
-    background-color: #ffffff;
-    width: 340px;
-    /* height: 275px; */
-    height: auto;
-    text-align: center;
-    position: fixed ;
-    display: none;
-    left: 0;
-    right: 0;
-    top: 80px;
-    margin: 0 auto;
-    border-radius: 5px;
-    border: none;
-    overflow: hidden;
-    z-index: 99;
-}
-#loginWindowBackground{
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    min-height: 100%;
-    background: rgba(0,0,0,0.6);
-    z-index: 88;
-    cursor: pointer;
-    display: none;
-}
-.registerCloser{
-    position: absolute;
-    top: 20px;
-    right: 10px;
-    width: 20px;
-    height: 20px;
-    cursor: pointer;
-}
-.registerBackground{
-    background: linear-gradient(270deg, #337788 0%, #66AABB 100%);
-    width: 100%;
-    height: 10px;
-}
-.registerTitle{
-    color: #666666;
-    font-size: 24px;
-    font-weight: bold;
-    height: 44px;
-    margin-top: 10px;
-}
-.registerBox{
-    width: 95%;
-    margin: 0 auto;
-}
-.messageOK{
-    color: green;
-}
-.registerBoxEmail, .registerBoxPass, .loginButton, .registerButton, .newUserName, .newEmail, .newPassword{
-    max-width: 310px;
-    width: 90%;
-    height: 22px;
-    border-radius: 5px;
-    border: 1px solid #CCCCCC;
-    margin: 5px auto;
-    padding: 10px;
-}
-.loginButton ,.registerButton{
-    background-color: #448899;
-    color: #ffffff;
-    max-width: 312px; /*310+1border+1border*/
-    width: 100%;
-    font-size: 17px;
-    height: 42px;
-    padding: none;
-    border: none;
-    cursor: pointer;
-}
-.registerFooter{
-    height: 44px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-.nullTop{
-    z-index: 1;
-    position: absolute;
-    height: 320px;
-    left: 0;
-    right: 0;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-.nullTop>div{
-    width: 1200px;
-    left: 1%;
-}
-.category{
-    z-index: 99;
-    position: relative;
-    left: 1%;
-}
-.listCatagory{
-    background-color: #ffffff;
-    position: absolute;
-    margin-top: 20px;
-    padding: 10px;
-    width: 400px;
-    display: none;
-    grid-template-columns: repeat(3,1fr);
-    border-radius: 5px;
-    box-shadow: 0px 0px 20px #AABBCC;
-}
-ul li{
-    list-style: none;
-    line-height: 33px;
-    cursor: pointer;
-}
-ul :hover{
-    background-color: #448899;
-    color: #ffffff;
-}
-.onWelcomePng{
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-.onWelcomePng>div{
-    color: #ffffff;
-    width: 1200px;
-    height: auto;
-    position: relative;
-    left: 1%;
-}
-#welcome{
-    height: 320px;
-    margin: 0 auto;
-    overflow: hidden;
-    position: relative;
- 
-}
-#welcome>img{
-    width: 100%;
-    overflow: hidden; 
-}
-
-/* searchBar */
-.categorySearch{
-    width: 500px; /*60px+400px*/
-    height: 30px;
-}
-.categorySearch>input[type=text] {
-    padding: 10px;
-    width: 80%;
-    font-size: 17px;
-    border: 1px solid grey;
-    float: left;
-    border-radius: 5px 0 0 5px;
-    background: #f1f1f1;
-    float: left;
-    cursor: pointer;
-
-}
-.categorySearch>button{
-    padding: 10px;
-    width: auto;
-    background: #448899;
-    color: white;
-    font-size: 17px;
-    border: 1px solid grey;
-    border-left: none;
-    cursor: pointer;
-    border-radius: 0 5px 5px 0;
-}
-
-/* main */
-#main{
-    width: 1200px;
-    margin: 20px auto;
-    display: grid;
-    grid-template-columns: repeat(4,1fr);
-    /* justify-content: space-between;    */
-    /* justify-content: space-evenly; */
-    /* justify-content: space-around; */
-    gap: 30px;
-}
-.mainImageBox {
-    width: auto;
-    height: 197px;
-    overflow: hidden;
-    cursor: pointer;
-}
-.mainImageBox :hover{
-    opacity: 0.4;
-}
-.mainImage{
-    width: 100%;
-    height: auto;  
-}
-.name-background{
-    background-color:rgba(0,0,0,0.6);
-    height: 40px;
-    width: 100%;
-    display: flex;
-    align-items: center;
-    position: absolute;
-    top: 157px; /* 197px-40px */
-}
-.name{
-    color: #ffffff;
-    margin: 4%;
-}
-.mrt{
-    float: left;
-    margin: 4%;
-}
-.cat{
-    float: right;
-    margin: 4%;
-}
-
-.item{
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    /* width: 270px; */
-    overflow: hidden;
-    position: relative;
-
-}
-footer{
-    color: #ffffff;
-    background-color: #757575;
-    height: 104px;
-    margin: 0 auto;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-
-@media (max-width:1200px){
-    #header{
-        width: 100%;
-    }
-    #welcome>img{
-        width: auto;
-        height: 320px;
-    }
-    #main{
-        width: 90%;
-        height: auto;
-        grid-template-columns: repeat(2,1fr);
-    }
-    
-    .mainImageBox{
-        height: 220px;
-    }
-    .name-background{
-        top: 180px;
-    }
-    footer{
-        width: 100%;
-    }
-}
-
-@media(max-width:600px){
-    .header-rightItem{
-        margin-left: auto;
-        line-height: 54px;
-    }
-    #main{
-        grid-template-columns: repeat(1,1fr);
-    }
-    .mainImageBox{
-        height: 235px;
-    }
-    .name-background{
-        top: 195px;
-    }
-    .categorySearch{
-        width: 90%;
-    }
-    
-}
-
-@media(max-width:360px){
-    #register{
-        width: 95%;
-    }
-    #login{
-        width: 95%;
-    }
-}
